FEAT: implement outbound embed functionality

+++ commands/utility/cerebrate.ts
@@ -61,14 +61,19 @@ export const command = {
 			embeds: [embed]
 		});
 
-		// TODO: Implement select menu for moving channel
+		let outRes: Message<true> | null = null;
+		let outBtnCollector: InteractionCollector<ButtonInteraction<'cached'>> | null = null; 
 
-		let outBtnCollector: InteractionCollector<ButtonInteraction<'cached'>> = res.createMessageComponentCollector({ componentType: ComponentType.Button, time: 3600000 });
+		let inBtnCollector: InteractionCollector<ButtonInteraction<'cached'>> = res.createMessageComponentCollector({ componentType: ComponentType.Button, time: 3600000 });
 
-		outBtnCollector.on('collect', async btn => {
+		inBtnCollector.on('collect', async btn => {
 			switch (btn.customId) {
 				case 'connect' : {
+					if (nexus.inboundCollector === null) {
+						throw new Error('Collector failed to initialize');
+					}
 					let reply = await btn.reply({ content: 'Joined.', ephemeral: true });
+					nexus.inboundCollector.on('collect', nexus.inCollectorFn);
 					setTimeout(() => reply.delete(), 5000);
 					break;
 				}
@@ -83,7 +88,6 @@ export const command = {
 					menuSelectCollector.on('collect', async selection => {
 						let selectedChannelId = selection.values[0];
 						let selectedChannel: TextChannel | undefined = selection.channels.get(selectedChannelId) as TextChannel;
-
 						// TODO: Check if sendable
 						if (!selectedChannel) return;
 
@@ -101,8 +105,10 @@ export const command = {
 							throw new Error('Outbound collecor not initialized');
 						}
 						targetChannel /*= nexus.outboundCollector.channel*/ = props.nexus.setChannelTarget(selectedChannel);
-						await nexus.redirectOutboundCollector(selectedChannel);
-						await nexus.redirectInboundCollector(selectedChannel);
+						// Reset collector
+						await nexus.redirectOutboundCollector(nexus.getSourceChannel());
+						// Actual redirect
+						await nexus.redirectInboundCollector({ newChannel: selectedChannel });
 						let notice = await selection.reply({ content: `Moved to <#${selectedChannelId}>\n-# Deleting <t:${Math.ceil(new Date().getTime() / 1000) + 5}:R>` });
 						menuSelectCollector.stop();
 						setTimeout(async () => {
@@ -112,9 +118,12 @@ export const command = {
 					});
 					break;
 				}
+				case 'leave': {
+					break;
+				}
 			}
 
-			if (outBtnCollector.collected.size === 1) {
+			if (inBtnCollector.collected.size === 1) {
 				let postConnectEmbed = await res.edit({
 					components: [
 						row1.setComponents(connect, leave, relocate),
@@ -134,15 +143,70 @@ export const command = {
 						.setTitle('Connection Established')],
 				});
 
-				let received = await (interaction.channel as TextChannel).send({
+				outRes = await (nexus.getSourceChannel() as TextChannel).send({
 					components: [row1.setComponents(connect, leave, relocate)],
 					embeds: [embed.setTitle('Connection Established').setDescription(null)]
 				});
 
-				let inBtnCollector = received.createMessageComponentCollector({ componentType: ComponentType.Button, time: 3600000 });
-				inBtnCollector.on('collect', async btn => {
-					let reply = await btn.reply({ content: 'Joined.', ephemeral: true });
-					setTimeout(() => reply.delete(), 5000);
+				outBtnCollector = outRes.createMessageComponentCollector({ componentType: ComponentType.Button, time: 3600000 });
+				outBtnCollector.on('collect', async btn => {
+					switch (btn.customId) {
+						case 'join': {
+							btn.channel?.send('You used \'Join\'');
+							let reply = await btn.reply({ content: 'Joined.', ephemeral: true });
+							setTimeout(() => reply.delete(), 5000);
+							break;
+						}
+						case 'relocate': {
+							let relocation = await btn.update({
+								components: [
+									row1.setComponents(connect, leave, relocate.setDisabled()),
+									row2,
+								],
+							})
+							let menuSelectCollector: InteractionCollector<ChannelSelectMenuInteraction<'cached'>> = relocation.createMessageComponentCollector({ componentType: ComponentType.ChannelSelect, time: 300000 });
+							menuSelectCollector.on('collect', async selection => {
+								let selectedChannelId = selection.values[0];
+								let selectedChannel: TextChannel | undefined = selection.channels.get(selectedChannelId) as TextChannel;
+								// TODO: Check if sendable
+								if (!selectedChannel) return;
+
+								let res = await relocation.fetch();
+
+								let relocatedEmbed = selectedChannel.send({ 
+									components: [row1.setComponents(connect, leave)],
+									embeds: res.embeds
+								});
+
+								if (selection.channel === null) return;
+
+								// Update nexus variables
+								if (nexus.outboundCollector === null) {
+									throw new Error('Outbound collecor not initialized');
+								}
+								// Actual redirect
+								let prevSourceChannel = nexus.getSourceChannel();
+								await nexus.redirectOutboundCollector(selectedChannel);
+								// Reset collector
+								let targetChannel = nexus.getChannelTarget();
+								if (targetChannel === null || prevSourceChannel === null) {
+									throw new Error('No channel target');
+								}
+								await nexus.redirectInboundCollector({ newChannel: selectedChannel, prevSourceChannel });
+								let notice = await selection.reply({ content: `Moved to <#${selectedChannelId}>\n-# Deleting <t:${Math.ceil(new Date().getTime() / 1000) + 5}:R>` });
+								menuSelectCollector.stop();
+								setTimeout(async () => {
+									await relocation.delete();
+									await notice.delete();
+								}, 5000);
+							});
+							break;
+						}
+						case 'leave': {
+							btn.channel?.send('You used \'Leave\'');
+							break;
+						}
+					}
 				});
 			}
 		});
